+++ src/main/java/net/silentchaos512/gear/util/Const.java
@@ -8,4 +8,8 @@ public class Const {
      * Example material ID
      */
     public static final ResourceLocation EXAMPLE = SilentGear.getId("example");
+
+    public static final ResourceLocation SALVAGING = SilentGear.getId("salvaging");
+    public static final ResourceLocation SALVAGING_COMPOUND_PART = SilentGear.getId("salvaging/compound_part");
+    public static final ResourceLocation SALVAGING_GEAR = SilentGear.getId("salvaging/gear");
 }